refactor(integration): own shared value objects in the domain layer
The domain imported IntegrationType from presentation and OAuthResource/SyncMetadata
from infrastructure, inverting the dependency direction. Move these (plus SkippedDetail)
into domain/value_objects; the domain entities/factories now import from there.
presentation/models and infrastructure/content_service/types re-export them for
backward compatibility, so the ~15 external importers are untouched and object
identity is preserved.
