+++ backend/src/intric/integration/domain/factories/sync_log_factory.py
@@ -2,7 +2,7 @@ from typing import cast
 
 from intric.database.tables.sync_log_table import SyncLog as SyncLogDBModel
 from intric.integration.domain.entities.sync_log import SyncLog
-from intric.integration.infrastructure.content_service.types import SyncMetadata
+from intric.integration.domain.value_objects import SyncMetadata
 
 
 class SyncLogFactory:
